requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
httpx==0.28.1
sqlalchemy==2.0.25
psycopg==3.1.18
python-dotenv==1.0.0
